+++ src/tagorganizer/widgets/tag_bar.py
@@ -18,6 +18,7 @@ along with TagOrganizer. If not, see <https://www.gnu.org/licenses/>.
 
 """
 
+from datetime import datetime
 from qtpy.QtWidgets import QHBoxLayout, QPushButton, QSizePolicy
 
 
@@ -26,7 +27,10 @@ class TagBar(QHBoxLayout):
         super().__init__()
 
         self.main = main
+
         self.selected_tags = []
+        self.selected_times_min = (None, None)
+        self.selected_times_max = (None, None)
 
         self.clear_button = QPushButton("Clear")
         self.clear_button.clicked.connect(self.clear_selected_tags)
@@ -45,18 +49,57 @@ class TagBar(QHBoxLayout):
     def add_tag(self, tag_name: str):
         tag_button = QPushButton(tag_name)
         tag_button.setSizePolicy(QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Minimum)
-        tag_button.clicked.connect(lambda: self.remove_tag_button(tag_button))
+        tag_button.clicked.connect(lambda flag, t=tag_button: self.remove_tag_button(t))
         self.addWidget(tag_button, 0)
         self.selected_tags.append((tag_name, tag_button))
+        self.main.update_items()
+
+    def add_time_tag(self, date: datetime, min_max: str = "<"):
+        # delete button if there is already one
+        if min_max == ">":
+            if self.selected_times_min[0] is not None:
+                w = self.selected_times_min[1]
+                w.setParent(None)
+                del w
+        elif min_max == "<":
+            if self.selected_times_max[0] is not None:
+                w = self.selected_times_max[1]
+                w.setParent(None)
+                del w
+
+        tag_name = f"{min_max} {date.strftime('%Y-%m-%d')}"
+        tag_button = QPushButton(tag_name)
+        tag_button.setSizePolicy(QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Minimum)
+        tag_button.clicked.connect(lambda flag, t=tag_button: self.remove_tag_button(t))
+        if min_max == ">":
+            self.selected_times_min = (date, tag_button)
+        elif min_max == "<":
+            self.selected_times_max = (date, tag_button)
+
+        self.addWidget(tag_button, 0)
+
         self.main.update_items()
 
     def get_selected_tags(self):
         return [x[0] for x in self.selected_tags]
 
     def remove_tag_button(self, w):
+        found = False
         for i, (_, widget) in enumerate(self.selected_tags):
             if w == widget:
+                found = True
                 break
-        w.setParent(None)
-        del self.selected_tags[i]
+        if found:
+            w.setParent(None)
+            del self.selected_tags[i]
+        else:
+            if w == self.selected_times_min[1]:
+                w.setParent(None)
+                self.selected_times_min = (None, None)
+                del w
+            elif w == self.selected_times_max[1]:
+                w.setParent(None)
+                self.selected_times_max = (None, None)
+                del w
+
         self.main.update_items()